BLACKWOOD MANOR
Level 1: 'Foundation of Fear' - Set in the woods of West Virginia, players take on the role of a detective investigating the disappearance of a group of friends. The level should feature a creepy atmosphere, with an emphasis on exploration and puzzle-solving.
Level 2: 'Dead End' - Inspired by Wrong Turn
 2: Dead End, this level takes place in a deserted town, where players must navigate through a maze of streets and buildings to escape the clutches of the cannibal mutants.

Level 3: 'Left for Dead' - Based on Wrong Turn 3: Left for Dead, this level is set in a prison, where players must fight for survival against the mutants and other prisoners.

Level 4: 'Bloody Beginnings' - This level is inspired by Wrong Turn 4: Bloody Beginnings, and takes place in a college town, where players must uncover the dark secrets behind a series of gruesome murders.

Level 5: 'Bloodlines' - Set in a rural town, this level is based on Wrong Turn 5: Bloodlines, and follows the story of a group of friends who are stalked and killed by the mutants.

Level 6: 'Last Resort' - The final level, inspired by Wrong Turn 6: Last Resort, takes place in a secluded resort, where players must confront the mastermind behind the mutants and put an end to the horror.

Throughout the game, incorporate jump scares, creepy sound effects, and a sense of unease to create a truly terrifying experience. The game should be designed to be played with headphones and in a dark room for maximum effect